Refactor plotting logic out of `SingleDiskPlot` into `plotting` module
